# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Integration API",
    "description": "Operations for API keys and service accounts. Authenticate with `Authorization: Bearer <credential>`; errors are RFC 9457 problem details; every response carries `x-request-id`. Examples: examples/api-clients.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/orgs/{slug}/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "The organization's audit trail, newest first.",
        "operationId": "audit_log",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Organization slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actor_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditRow"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or revoked credential. Never falls back to another method.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "The credential's scopes do not include the operation's permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Unknown resource, or an organization the credential does not belong to.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited; wait for `Retry-After` seconds.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": [
              "audit:read"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "description": "Field-level validation problem.",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "Page_AuditRow": {
        "type": "object",
        "description": "A page of results plus the cursor for the next page (None when exhausted).",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "occurred_at",
                "actor_type",
                "action",
                "outcome",
                "metadata"
              ],
              "properties": {
                "action": {
                  "type": "string"
                },
                "actor_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "actor_label": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "actor_type": {
                  "type": "string"
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "metadata": {},
                "occurred_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "organization_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "outcome": {
                  "type": "string"
                },
                "request_id": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "target_id": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "target_type": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem document.",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "An organization API key, or a service account's JWT from the identity provider."
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}